+++ src/main/java/com/solo/todolist2/repository/TodosRepository.java
@@ -4,5 +4,5 @@ import com.solo.todolist2.entity.Todos;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TodosRepository extends JpaRepository<Todos, Integer> {
+public interface TodosRepository extends JpaRepository<Todos, Long> {
 }
